refactor(insights): hoist BarChart to shared tabs/components
